Implement various enhancements and fixes across multiple files
- Added checks to prevent deletion and disabling of captchas if only one remains in the database.
- Improved user feedback for volunteer candidates and unban requests when no data is available.
- Added a notification system for new chat messages, including permission requests for notifications.
- Implemented a delete message feature with appropriate checks for ownership.
- Enhanced petition creation with stricter validation and character filtering.
- Updated user interface to handle cases where no petitions or signatures exist.
- Added error handling for blocked users in chat functionality.
- Improved CSS styles for better visual feedback and layout adjustments.
- Updated database schema to include QR code filename generation for petitions.

+++ Sources/BackOffice/Processus/delete_captcha.php
@@ -8,6 +8,16 @@ if($is_admin != 0){
         $id = intval($_GET['id']);
 
         try {
+
+            $count_how_many_captchas_are_in_database = $pdo->prepare("SELECT COUNT(*) FROM CAPTCHA WHERE state = 1");
+            $count_how_many_captchas_are_in_database->execute();
+            $count = $count_how_many_captchas_are_in_database->fetchColumn();
+
+            if ($count == 1) {
+                echo "<script>window.location.href = '../captcha.php?error=TooFew';</script>";
+                exit();
+            }
+
             $stmt = $pdo->prepare("DELETE FROM CAPTCHA WHERE id = :id");
             $stmt->bindParam(':id', $id, PDO::PARAM_INT);
